bugfix: Relax argument checking of HANDLE to allow loading of the null handle.
This patch relaxes the argument checking of the HANDLE bytecode, to allow
loading of the NULL handle. The original version of the code handed out
in the assignment specificetion (r1.0) was a bit more restrictive than
initially intended.

With this patch, it now is possible to use "HANDLE(NULL)" as originally
intended. (Note: Previously it would have been possible to load the NULL
handle by using a plain "IMM(0)")